+++ sail/csrc/core/autograd/UnaryGrad.cpp
@@ -1,4 +1,4 @@
-#include "unary_function.h"
+#include "UnaryGrad.h"
 #include <iostream>
 #include <vector>
 #include "Tensor.h"
@@ -12,21 +12,6 @@ namespace sail {
 namespace autograd {
 
 using TensorVector = std::vector<Tensor>;
-
-Tensor Exp::forward(TensorVector inputs) { return ops::exp(inputs[0]); }
-TensorVector Exp::backward(Tensor& grad) {
-    Tensor t = ops::exp(Function::arg_storage[0]);
-    Tensor b = t * grad;
-    return {b};
-}
-
-Tensor Log::forward(TensorVector inputs) {
-    stored_log = ops::log(inputs[0]);
-    return stored_log;
-}
-TensorVector Log::backward(Tensor& grad) {
-    return {grad / Function::arg_storage[0]};
-}
 
 Tensor ClipMinOnly::forward(TensorVector inputs) {
     return ops::clip_min(inputs[0], min);
