+++ sail/csrc/core/autograd/UnaryGrad.h
@@ -13,21 +13,6 @@ namespace autograd {
 
 using TensorVector = std::vector<Tensor>;
 
-class Exp : public Function {
-   public:
-    explicit Exp(){};
-    Tensor forward(TensorVector inputs);
-    TensorVector backward(Tensor& grad);
-};
-
-class Log : public Function {
-   public:
-    Tensor stored_log;
-    explicit Log(){};
-    Tensor forward(TensorVector inputs);
-    TensorVector backward(Tensor& grad);
-};
-
 class ClipMinOnly : public Function {
    public:
     Tensor stored_log;
